+++ Libplanet.Net/Messages/ConsensusMessage.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections.Generic;
 using Libplanet.Blocks;
+using Libplanet.Crypto;
 using Libplanet.Net.Consensus;
 
 namespace Libplanet.Net.Messages
@@ -15,11 +16,18 @@ namespace Libplanet.Net.Messages
         /// <summary>
         /// Initializes a new instance of the <see cref="ConsensusMessage"/> class.
         /// </summary>
+        /// <param name="validator">
+        /// A <see cref="PublicKey"/> of the validator who made this message.</param>
         /// <param name="height">A <see cref="Context{T}.Height"/> the message is for.</param>
         /// <param name="round">A <see cref="Context{T}.Round"/> the message is written for.</param>
         /// <param name="blockHash">A <see cref="BlockHash"/> the message is written for.</param>
-        protected ConsensusMessage(long height, int round, BlockHash? blockHash)
+        protected ConsensusMessage(
+            PublicKey validator,
+            long height,
+            int round,
+            BlockHash? blockHash)
         {
+            Validator = validator;
             Round = round;
             Height = height;
             BlockHash = blockHash;
@@ -32,17 +40,23 @@ namespace Libplanet.Net.Messages
         /// <param name="dataframes">A marshalled message.</param>
         protected ConsensusMessage(byte[][] dataframes)
         {
-            Height = BitConverter.ToInt64(dataframes[0], 0);
-            Round = BitConverter.ToInt32(dataframes[1], 0);
-            if (dataframes[2].Length == 1 && dataframes[2][0] == Nil)
+            Validator = new PublicKey(dataframes[0]);
+            Height = BitConverter.ToInt64(dataframes[1], 0);
+            Round = BitConverter.ToInt32(dataframes[2], 0);
+            if (dataframes[3].Length == 1 && dataframes[3][0] == Nil)
             {
                 BlockHash = null;
             }
             else
             {
-                BlockHash = new BlockHash(dataframes[2]);
+                BlockHash = new BlockHash(dataframes[3]);
             }
         }
+
+        /// <summary>
+        /// A <see cref="PublicKey"/> of the validator who made this message.
+        /// </summary>
+        public PublicKey Validator { get; }
 
         /// <summary>
         /// A <see cref="Context{T}.Height"/> the message is written for.
@@ -64,6 +78,7 @@ namespace Libplanet.Net.Messages
         /// </summary>
         public override IEnumerable<byte[]> DataFrames => new[]
         {
+            Validator.Format(true),
             BitConverter.GetBytes(Height),
             BitConverter.GetBytes(Round),
             BlockHash is { } blockHash ? blockHash.ToByteArray() : new[] { Nil },