+++ CHANGES.md
@@ -48,9 +48,8 @@ To be released.
  -  (Libplanet.Net) Added `InvalidHeightIncreasingException` class.  [[#PBFT]]
  -  (Libplanet.Net) Added
     `InvalidProposerProposeMessageException` class.  [[#PBFT]]
- -  (Libplanet.Net) Added
-    `InvalidValidatorVoteMessageException` class.  [[#PBFT]]
- -  (Libplanet.Net) Added `Message.GetHash()` method.  [[#PBFT]]
+ -  (Libplanet.Net) Added `Message.Id` property.  [[#PBFT]]
+ -  (Libplanet.Net) Added `Gossip` class.  [[#PBFT]]
 
 ### Behavioral changes